# Apply Various 1D Fabric Optimizations - Improve Performance by ~500 MB/s for 4k packet size