--- a/ttnn/cpp/ttnn/operations/ccl/kernels/edm_fabric/fabric_erisc_datamover.cpp
+++ b/ttnn/cpp/ttnn/operations/ccl/kernels/edm_fabric/fabric_erisc_datamover.cpp
@@ -657,6 +657,7 @@ FORCE_INLINE void receiver_forward_packet(
             default:
                 ASSERT(false);
         }
+    }
 }
 
 ////////////////////////////////////